Use InlineButton style for Tags
